Added TrackPage

+++ src/main/java/com/example/timetrack/entity/Track.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -32,5 +33,5 @@ public class Track {
 
     @NotNull
     @Column(name = "time")
-    private Double time;
+    private BigDecimal time;
 }
